third_party/WebKit/LayoutTests/fast/text/first-letter-bad-line-boxes-crash-expected.txt

CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
CONSOLE WARNING: We don't execute document.execCommand() this time, because it is called recursively.
CONSOLE ERROR: line 66: Uncaught NotFoundError: Failed to execute 'surroundContents' on 'Range': The node provided is null.
\\\\\\m{VVVVVVVV\\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW
 \\\\\\m{VVVVVVVVVVVVVVV<};<WWWWW
xt
